move "centroid" to "location"
