+++ specutils/analysis/__init__.py
@@ -2,5 +2,5 @@ from __future__ import absolute_import
 
 from .flux import *  # noqa
 from .snr import *  # noqa
-from .centroid import *  # noqa
+from .location import *  # noqa
 from .width import *  # noqa
