+++ specutils/analysis/location.py
